SF: Correctly handle the LayerRequest bitfield
As LayerRequest is a bitfield ('1 << 0' is used in the enum
declaration) we need to AND the bit individually rather than the whole
value for each flag, furthermore to_string cannot be used for unknown
values as it only prints values already in the enum, which logically
wouldn't be unknown. This change is needed on Nvidia devices to avoid
missing the ClearClient target request when it's submitted together with
their custom ones.

Change-Id: I244c50488fbfc4d1b93c0f9409bed15c39668c76
Signed-off-by: Ray <vincentrayananda@gmail.com>